requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
PyMySQL==1.1.1
cryptography==44.0.0
openpyxl==3.1.5
google-generativeai==0.8.4
Werkzeug==3.1.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
